requirements.txt
Django>=2.2.6
djangorestframework>=3.10.3
numpy>=1.17.2
pytz>=2019.2
sqlparse>=0.3.0
waitress>=1.3.1
whitenoise>=4.1.4
